Add files via upload
The current COVID-19 pandemic has created several problems nowadays.
All these problems point to the inability to examine and scale the situation according to the intensity of
the outbreak. Along with creating problems, this pandemic has exceeded many boundaries such as the provincial, radical, conceptual, spiritual, social, and educational. It has changed the fundamental nature of our society.
There is one answer to solve this problem and the answer is to utilize the technology in the best way. 
For controlling such a pandemic situation, a collection, analysis, elucidation of data on regular basis regarding the spreading disease trends should be done in order to predict the outbreak of major health-related symptoms. Proper surveillance is required to cure any type of disease.
Surveillance- a collection of data can be performed best with the assistance of technology like the Internet of 
things.
As “Internet of things is an idea which comprises of interconnected frameworks which have the ability 
of finding data about a specific thing with the assistance of extraordinary identifiers and detecting 
capability.”
The technological platform of IoT connected with the healthcare system is helpful in observing the infected people. This monitoring is done by the interconnected network offered by IoT. [1]
Implementation of such a technology will help to decrease healthcare expenses and enhance the treatment of the infected patients.
Therefore, this study is an extensive description about exploring and highlighting the prevalence of IoT in 
pandemic.