Add files via upload
The prime motive of the emerging technology named the Internet of things (IoT) is to bring everything including wearable sensors, medical care sensors, cameras and home machines, advanced mobile phones, and so on the web or internet. These various things create enormous information which subsequently leads to the need for prerequisites of productive stockpiling and preparing. Fog computing is an arising innovation to defeat this issue. Nonetheless, there are a few applications (medical care) that need to deal with information that continues to improve its presentation and require low idleness and postponement. Fog computing is one of the efficient solutions for providing better facilities in the domain of medical care. In this research, a fog-based framework is presented for distant wellbeing monitoring. 
This proposed framework majorly focuses on monitoring the blood glucose level of patients suffering from diabetes and having cardiovascular disease. It also ensures to limit energy utilization. 
Exploratory outcomes uncover that the presentation of the proposed structure is productive regarding 
network postponement and energy utilization.